add pg_build_env script and remove unused base class
