+++ src/sub_platforms/sql_opt/column_statastics/statistics_info.py
@@ -3,12 +3,12 @@ Copyright (c) 2024 Bytedance Ltd. and/or its affiliates
 SPDX-License-Identifier: MIT
 """
 from typing import Dict, List, Any, Optional
-from pydantic import Field, PrivateAttr, PlainSerializer, BeforeValidator
+from pydantic import BaseModel, Field, PrivateAttr, BeforeValidator
 from typing_extensions import Annotated
 import pandas as pd
 
+from sub_platforms.sql_opt.common.pydantic_utils import PydanticDataClassJsonMixin
 from sub_platforms.sql_opt.videx.videx_histogram import HistogramStats
-from sub_platforms.sql_opt.column_statastics.statistics_info_base import BaseTableStatisticsInfo
 
 def large_number_decoder(y):
     if isinstance(y, list):
@@ -21,7 +21,7 @@ def large_number_decoder(y):
         return res
 
 
-class TableStatisticsInfo(BaseTableStatisticsInfo):
+class TableStatisticsInfo(BaseModel, PydanticDataClassJsonMixin):
     model_config = {"arbitrary_types_allowed": True}
     db_name: str
     table_name: str
