--- src/sub_platforms/sql_opt/column_statastics/statistics_info_base.py
@@ -1,5 +1,0 @@
-from pydantic import BaseModel
-from sub_platforms.sql_opt.common.pydantic_utils import PydanticDataClassJsonMixin
-
-class BaseTableStatisticsInfo(BaseModel, PydanticDataClassJsonMixin):
-    pass
