+++ src/sub_platforms/sql_opt/column_statastics/statistics_info_pg.py
@@ -1,8 +1,9 @@
 from typing import Dict, Optional
-from pydantic import Field
+from pydantic import BaseModel, Field
+from sub_platforms.sql_opt.common.pydantic_utils import PydanticDataClassJsonMixin
 from sub_platforms.sql_opt.pg_meta import PGStatistic,PGStatisticExt
-from sub_platforms.sql_opt.column_statastics.statistics_info_base import BaseTableStatisticsInfo
-class PGTableStatisticsInfo(BaseTableStatisticsInfo):
+
+class PGTableStatisticsInfo(BaseModel, PydanticDataClassJsonMixin):
     """ PostgreSQL Table Statistics Information """
     db_name: str
     schema_name: str
