+++ src/sub_platforms/sql_opt/databases/pg/explain_result.py
@@ -1,6 +1,4 @@
-import pandas as pd
-from typing import List, Optional, Union
-from pydantic import BaseModel, Field
+from pydantic import BaseModel
 
 from sub_platforms.sql_opt.common.pydantic_utils import PydanticDataClassJsonMixin
